Create version_control plugin directory and move gitlab into it
