+++ nornir/plugins/tasks/files/__init__.py
@@ -1,6 +1,4 @@
 from .sftp import sftp
 from .write_file import write_file
-from .gitlab import gitlab
 
-
-__all__ = ("sftp", "write_file", "gitlab")
+__all__ = ("sftp", "write_file")
