+++ nornir/plugins/tasks/version_control/__init__.py
@@ -1,0 +1,3 @@
+from .gitlab import gitlab
+
+__all__ = ("gitlab",)
