+++ nornir/plugins/tasks/version_control/gitlab.py
